AIReels/requirements.txt
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
ffmpeg-python==0.2.0 